Ra-vision-ai/requirements.txt
fastapi
uvicorn
langchain-huggingface
langchain-core
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
